another commit

+++ app/src/main/java/com/konradpekala/hackyeah2018/data/model/CardInfo.kt
@@ -1,6 +1,6 @@
 package com.konradpekala.hackyeah2018.data.model
 
-data class CardInfo(val cardNumber: Int,
+data class CardInfo(val cardNumber: Long,
                     val expirationYear: Int,
                     val expirationMonth: Int,
                     val cvc: Int)
